hacked-it/requirements.txt
pillow 
rich